commiting changes

+++ musicapp/api/song.py
@@ -1,4 +1,5 @@
 from flask import jsonify, make_response, request
+from flask_login import current_user
 from musicapp.api import api_views
 from musicapp import database
 from musicapp.models.song import Song
@@ -21,13 +22,15 @@ def remove_song(id):
     if not id:
         return make_response(jsonify({"error": True, "message" : "Bad request"}), 400)
     try:
-
-        song = Song.query.filter_by(id=id).one()
-        database.session.delete(song)
-        database.session.commit()
-
-        ret = {"success": True, "message": "song deleted successfully"}
-        return make_response(jsonify(ret), 200)
+        song = Song.query.filter_by(id=id and user_id==current_user.id).one()
+        if (song):
+            database.session.delete(song)
+            database.session.commit()
+            
+            ret = {"success": True, "message": "song deleted successfully"}
+            return make_response(jsonify(ret), 200)
+        else:
+            raise NoResultFound("you are not allowed")
     except NoResultFound as e:
         ret_err = {"error" : True, "message" : "Bad request"}
         return make_response(jsonify(ret_err), 400)
